Add a GUI interface with the Tk toolkit
Tk is shipped with Python distribution. For a end user, it's much
easier to start using GUI applications writing in the ugly Tk toolkit.

Since Tk has an awful API for a GUI toolkit, there will not be too
much features packed in the interface.

A new command is provided to start the GUI interface:
  $ ./you-get-tk